+++ db/migrate/20120531103056_create_order_items.rb
@@ -5,11 +5,11 @@ class CreateOrderItems < ActiveRecord::Migration
       t.integer :sku_id,        :null => false
       t.string  :type,          :null => false, :limit => 35, :default => 'OrderRegularItem'
       t.integer :quantity,      :null => false, :limit => 3
-      t.integer :actual_price,  :null => false, :precision => 7,  :scale => 2
-      t.integer :actual_total,  :null => false, :precision => 7,  :scale => 2
-      t.integer :discount,      :null => false, :precision => 7,  :scale => 2, :default => 0
-      t.integer :price,         :null => false, :precision => 7,  :scale => 2
-      t.integer :total,         :null => false, :precision => 7,  :scale => 2
+      t.float   :actual_price,  :null => false, :precision => 7,  :scale => 2
+      t.float   :actual_total,  :null => false, :precision => 7,  :scale => 2
+      t.float   :discount,      :null => false, :precision => 7,  :scale => 2, :default => 0
+      t.float   :price,         :null => false, :precision => 7,  :scale => 2
+      t.float   :total,         :null => false, :precision => 7,  :scale => 2
 
       t.timestamps
     end
